LinearHypothesisTest

GitHub Repo: https://github.com/eohne/LinearHypothesisTest.jl

Perform Linear Hypothesis Tests. Currently allows to test hypothesis of the folling form for GLM and FixedEffectModels models:

math \beta_1 = \beta_2

math \beta_2 + 3 \times \beta_2 =5

math 0.5 \times \beta_1 + \beta_2 + beta_3 = \beta_4

Package Installation

The package is NOT registered with the General Registry but you can install the package either by using the Pkg REPL mode (press ] to enter):

pkg> add http://github.com/eohne/LinearHypothesisTest.jl

or by using Pkg functions

julia> using Pkg; Pkg.add("http://github.com/eohne/LinearHypothesisTest.jl")

Function Definition

LinearHypothesisTest.LinearHypothesisTests — Function

LinearHypothesisTests(lhs::Vector{<:Number},rhs::Number,coef::Vector{<:Number},vcov::Matrix{Float64},dof=nothing;test="")

Perform linear hypothesis tests on estimated coefficents. Can perform either an F or a Chisquared test. Returns a NamedTuple with the standard error, the Chisquared/F value, a p-value, the type of test performed, and the 5% critical value.

Args:

lhs::Vector:This corresponds to the hypothesis matrix. If you want to test whether b2 + 2* b3 = b4 and you have 5 coefficents from b1 to b5 (note the intercept counts as one of these coefficents) you would specify it as [0, 1, 2, -1 0]

rhs::Number: The value that the above linear equation should equal to under the null.

coef::Vector{Number}: A vector of estimates for (b1, b2, b3 etc)

vcov::Matrix{Float64}: A variance covariance matrix.

dof: The residual degrees of freedom. If these are not supplied a Chisquared Test will be performed.

test: One of "F" for an F-test or "C" for a Chisquared test.